+++ src/main/java/tn/esprit/spring/entity/Questionnaires.java
@@ -16,13 +16,13 @@ import javax.persistence.Table;
 public class Questionnaires implements Serializable{
 	@Id
 	@GeneratedValue (strategy = GenerationType.IDENTITY)
-	private int id; // Clé primaire
+	private Long id; // Clé primaire
 	
 	private String name;
 	private String description;
 	
-	public int getId() {return id;}
-	public void setId(int id) {this.id = id;}
+	public Long getId() {return id;}
+	public void setId(Long id) {this.id = id;}
 	public String getName() {return name;}
 	public void setName(String name) {this.name = name;}
 	public String getDescription() {return description;}
